Logik in Kippen zum Laufen gebracht Fehler behoben: Portraitorientierung bei Rückkehr zur Startseite war nicht gespeicherter Wert eindeutigere Benennung von Variablen der Startseite und auslagen der String in strings.xml Rückkehr zur Startseite bei Kippen verbessert Erinnerung an Irdroid bei Rückkehr zur Startseite ausgeschalten

+++ finaleApp/AirSwimmer/src/de/airswimmer/gui/Activity_Tilt_Permanent.java
@@ -24,7 +24,7 @@ public class Activity_Tilt_Permanent extends BaseTilt {
 	}
 
 	// method for movement of the fish
-	public void move(int xAxis, int yAxis) {
+	public void move(float xAxis, float yAxis) {
 		// TODO: implement movement
 	}
 }
